refactor: split PromptService into focused sub-services
PromptService had grown into a god object (866 lines) mixing four
concerns: low-level DIAL Core prompt-resource I/O, personal-bucket
prompt CRUD/listing, organisation/public-bucket reads, and folder
lifecycle + moves. Splits it into PromptsResourceService (shared
resource I/O, injected by the other three), PromptsPersonalService,
PromptsPublicService, and PromptsFolderService, plus a thin
delegation facade -- following the same facade + bound-property
pattern used for ConversationService and DeploymentsService/
ToolsetsService.

No REST contract, DTO, or behavior changes -- verified via the
relocated test suites (1945/1945 chat-api tests pass) plus lint/build.
