+++ apps/chat-api/src/prompts/prompt.module.ts
@@ -1,9 +1,19 @@
 import { Module } from '@nestjs/common';
+import { PromptsFolderService } from './folder/prompts-folder.service';
+import { PromptsPersonalService } from './personal/prompts-personal.service';
 import { PromptController } from './prompt.controller';
 import { PromptService } from './prompt.service';
+import { PromptsPublicService } from './public/prompts-public.service';
+import { PromptsResourceService } from './resource/prompts-resource.service';
 
 @Module({
   controllers: [PromptController],
-  providers: [PromptService],
+  providers: [
+    PromptService,
+    PromptsResourceService,
+    PromptsPersonalService,
+    PromptsPublicService,
+    PromptsFolderService,
+  ],
 })
 export class PromptModule {}
