+++ openspec/changes/archive/2026-08-10-split-prompt-service/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-08-10
